Split tests into different files for readability
