+++ tests/test_dump.py
@@ -1,21 +1,10 @@
 import pytest
-from jsonschema import Draft4Validator
 from marshmallow import Schema, fields, validate
 
 from marshmallow_jsonschema import JSONSchema
-from marshmallow_jsonschema.base import UnsupportedValueError, INCLUDE, EXCLUDE, RAISE
-from marshmallow_jsonschema.compat import (
-    dot_data_backwards_compatible,
-    MARSHMALLOW_MAJOR_VERSION,
-)
-from . import UserSchema, Address
-
-
-def _validate_schema(schema):
-    """
-    raises jsonschema.exceptions.SchemaError
-    """
-    Draft4Validator.check_schema(schema)
+from marshmallow_jsonschema.base import UnsupportedValueError
+from marshmallow_jsonschema.compat import dot_data_backwards_compatible
+from . import _validate_schema, UserSchema
 
 
 def test_dump_schema():
@@ -347,142 +336,6 @@ def test_nested_recursive():
     assert "RecursiveSchema" in props["children"]["items"]["$ref"]
 
 
-def test_one_of_validator():
-    schema = UserSchema()
-    json_schema = JSONSchema()
-
-    dumped = dot_data_backwards_compatible(json_schema.dump(schema))
-
-    _validate_schema(dumped)
-
-    assert dumped["definitions"]["UserSchema"]["properties"]["sex"]["enum"] == [
-        "male",
-        "female",
-        "non_binary",
-        "other",
-    ]
-    assert dumped["definitions"]["UserSchema"]["properties"]["sex"]["enumNames"] == [
-        "Male",
-        "Female",
-        "Non-binary/fluid",
-        "Other",
-    ]
-
-
-def test_range_validator():
-    schema = Address()
-    json_schema = JSONSchema()
-
-    dumped = dot_data_backwards_compatible(json_schema.dump(schema))
-
-    _validate_schema(dumped)
-
-    props = dumped["definitions"]["Address"]["properties"]
-    assert props["floor"]["minimum"] == 1
-    assert props["floor"]["maximum"] == 4
-
-
-def test_length_validator():
-    schema = UserSchema()
-    json_schema = JSONSchema()
-
-    dumped = dot_data_backwards_compatible(json_schema.dump(schema))
-
-    _validate_schema(dumped)
-
-    props = dumped["definitions"]["UserSchema"]["properties"]
-    assert props["name"]["minLength"] == 1
-    assert props["name"]["maxLength"] == 255
-    assert props["addresses"]["minItems"] == 1
-    assert props["addresses"]["maxItems"] == 3
-    assert props["const"]["minLength"] == 50
-    assert props["const"]["maxLength"] == 50
-
-
-def test_length_validator_value_error():
-    class BadSchema(Schema):
-        bob = fields.Integer(validate=validate.Length(min=1, max=3))
-
-        class Meta:
-            strict = True
-
-    schema = BadSchema()
-    json_schema = JSONSchema()
-
-    with pytest.raises(ValueError):
-        json_schema.dump(schema)
-
-
-def test_handle_range_not_number_returns_same_instance():
-    class SchemaWithStringRange(Schema):
-        floor = fields.String(validate=validate.Range(min=1, max=4))
-
-        class Meta:
-            strict = True
-
-    class SchemaWithNoRange(Schema):
-        floor = fields.String()
-
-        class Meta:
-            strict = True
-
-    class SchemaWithIntRangeValidate(Schema):
-        floor = fields.Integer(validate=validate.Range(min=1, max=4))
-
-        class Meta:
-            strict = True
-
-    class SchemaWithIntRangeNoValidate(Schema):
-        floor = fields.Integer()
-
-        class Meta:
-            strict = True
-
-    schema1 = SchemaWithStringRange()
-    schema2 = SchemaWithNoRange()
-    schema3 = SchemaWithIntRangeValidate()
-    schema4 = SchemaWithIntRangeNoValidate()
-    json_schema = JSONSchema()
-
-    # Delete "$ref" as root object names will obviously differ for schemas with different names
-    dumped_1 = dot_data_backwards_compatible(json_schema.dump(schema1))
-    del dumped_1["$ref"]
-    dumped_2 = dot_data_backwards_compatible(json_schema.dump(schema2))
-    del dumped_2["$ref"]
-
-    assert dumped_1 == dumped_2
-    assert json_schema.dump(schema3) != json_schema.dump(schema4)
-
-
-def test_handle_range_no_minimum():
-    class SchemaMin(Schema):
-        floor = fields.Integer(validate=validate.Range(min=1, max=4))
-
-        class Meta:
-            strict = True
-
-    class SchemaNoMin(Schema):
-        floor = fields.Integer(validate=validate.Range(max=4))
-
-        class Meta:
-            strict = True
-
-    schema1 = SchemaMin()
-    schema2 = SchemaNoMin()
-    json_schema = JSONSchema()
-
-    dumped1 = dot_data_backwards_compatible(json_schema.dump(schema1))["definitions"][
-        "SchemaMin"
-    ]
-    dumped2 = dot_data_backwards_compatible(json_schema.dump(schema2))["definitions"][
-        "SchemaNoMin"
-    ]
-    assert dumped1["properties"]["floor"]["minimum"] == 1
-    assert "exclusiveMinimum" not in dumped1["properties"]["floor"].keys()
-    assert "minimum" not in dumped2["properties"]["floor"]
-    assert "exclusiveMinimum" not in dumped2["properties"]["floor"]
-
-
 def test_title():
     class TestSchema(Schema):
         myfield = fields.String(metadata={"title": "Brown Cowzz"})
@@ -610,121 +463,3 @@ def test_required_excluded_when_empty():
     dumped = dot_data_backwards_compatible(json_schema.dump(schema))
 
     assert "required" not in dumped["definitions"]["TestSchema"]
-
-
-def test_additional_properties_default():
-    class TestSchema(Schema):
-        foo = fields.Integer()
-
-    schema = TestSchema()
-    json_schema = JSONSchema()
-
-    dumped = dot_data_backwards_compatible(json_schema.dump(schema))
-
-    assert not dumped["definitions"]["TestSchema"]["additionalProperties"]
-
-
-@pytest.mark.parametrize("additional_properties_value", (False, True))
-def test_additional_properties_from_meta(additional_properties_value):
-    class TestSchema(Schema):
-        class Meta:
-            additional_properties = additional_properties_value
-
-        foo = fields.Integer()
-
-    schema = TestSchema()
-    json_schema = JSONSchema()
-
-    dumped = dot_data_backwards_compatible(json_schema.dump(schema))
-
-    assert (
-        dumped["definitions"]["TestSchema"]["additionalProperties"]
-        == additional_properties_value
-    )
-
-
-def test_additional_properties_invalid_value():
-    class TestSchema(Schema):
-        class Meta:
-            additional_properties = "foo"
-
-        foo = fields.Integer()
-
-    schema = TestSchema()
-    json_schema = JSONSchema()
-
-    with pytest.raises(UnsupportedValueError):
-        json_schema.dump(schema)
-
-
-def test_additional_properties_nested_default():
-    class TestNestedSchema(Schema):
-        foo = fields.Integer()
-
-    class TestSchema(Schema):
-        nested = fields.Nested(TestNestedSchema())
-
-    schema = TestSchema()
-    json_schema = JSONSchema()
-
-    dumped = dot_data_backwards_compatible(json_schema.dump(schema))
-
-    assert not dumped["definitions"]["TestSchema"]["additionalProperties"]
-
-
-@pytest.mark.parametrize("additional_properties_value", (False, True))
-def test_additional_properties_from_nested_meta(additional_properties_value):
-    class TestNestedSchema(Schema):
-        class Meta:
-            additional_properties = additional_properties_value
-
-        foo = fields.Integer()
-
-    class TestSchema(Schema):
-        nested = fields.Nested(TestNestedSchema())
-
-    schema = TestSchema()
-    json_schema = JSONSchema()
-
-    dumped = dot_data_backwards_compatible(json_schema.dump(schema))
-
-    assert (
-        dumped["definitions"]["TestNestedSchema"]["additionalProperties"]
-        == additional_properties_value
-    )
-
-
-@pytest.mark.parametrize(
-    "unknown_value, additional_properties",
-    ((RAISE, False), (INCLUDE, True), (EXCLUDE, False)),
-)
-def test_additional_properties_deduced(unknown_value, additional_properties):
-    class TestSchema(Schema):
-        class Meta:
-            unknown = unknown_value
-
-        foo = fields.Integer()
-
-    schema = TestSchema()
-    json_schema = JSONSchema()
-
-    dumped = dot_data_backwards_compatible(json_schema.dump(schema))
-
-    assert (
-        dumped["definitions"]["TestSchema"]["additionalProperties"]
-        == additional_properties
-    )
-
-
-def test_additional_properties_unknown_invalid_value():
-    class TestSchema(Schema):
-        class Meta:
-            unknown = "foo"
-
-        foo = fields.Integer()
-
-    schema = TestSchema()
-    json_schema = JSONSchema()
-
-    with pytest.raises(UnsupportedValueError):
-        json_schema.dump(schema)
